telegram-bot-project/src/utils/embeddings.py
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import normalize
import numpy as np

def generate_embeddings(texts):
    # Placeholder for embedding generation logic
    # This function should convert a list of texts into their corresponding embeddings
    # For example, using a pre-trained model like Sentence Transformers
    embeddings = np.random.rand(len(texts), 768)  # Example: 768-dimensional embeddings
    return normalize(embeddings)

def calculate_similarity(embedding1, embedding2):
    # Calculate cosine similarity between two embeddings
    return cosine_similarity([embedding1], [embedding2])[0][0]

def find_most_similar(embedding, embeddings):
    # Find the index of the most similar embedding
    similarities = cosine_similarity([embedding], embeddings)
    return np.argmax(similarities)