telegram-bot-project/requirements.txt
FastAPI
SQLAlchemy
langchain
scikit-learn
pydantic
python-dotenv
uvicorn
numpy
requests